There is an issue with the getting the light vector for the quad I have just found
Because a quad is made up of four points unlike the sphere or box the light vector is struggling to work based of the position of the object because it doesnt know it, I have tried using the centre of the quad but it does not appear to make a diffrence. There should be some way to calculate it properly but I cant seem to find it.